merge

+++ src/main/java/com/youprice/onion/dto/member/BlockListDTO.java
@@ -15,7 +15,7 @@ public class BlockListDTO {
     private Long targetId;
     private String memberNickname;
     private String memberImageName;
-    private int userGrade;
+    private Double userGrade;
     private int complaintCount;
     private int blockListSize;
